able to fetch whole file

+++ src/peer_messages.rs
@@ -70,14 +70,14 @@ impl PeerMessage {
 
 #[derive(Debug)]
 pub struct PiecePayload {
-    pub index: usize,
+    pub index: u32,
     pub begin: u32,
     pub block: Vec<u8>,
 }
 
 impl PiecePayload {
     pub fn from_bytes(bytes: &[u8]) -> Self {
-        let index = usize::from_be_bytes(bytes[0..4].try_into().expect("Failed to convert index"));
+        let index = u32::from_be_bytes(bytes[0..4].try_into().expect("Failed to convert index"));
         let begin = u32::from_be_bytes(bytes[4..8].try_into().expect("Failed to convert begin"));
         let block = bytes[8..].to_vec();
         Self {
@@ -90,13 +90,13 @@ impl PiecePayload {
 
 #[derive(Debug)]
 pub struct RequestPayload {
-    pub index: usize,
+    pub index: u32,
     pub begin: u32,
     pub length: u32,
 }
 
 impl RequestPayload {
-    pub fn new(index: usize, begin: u32, length: u32) -> Self {
+    pub fn new(index: u32, begin: u32, length: u32) -> Self {
         Self {
             index,
             begin,
